requirements.txt
basicsr~=1.4.2
diffusers~=0.3.0
fairscale~=0.4.9
fonts~=0.0.3
font-roboto
gfpgan~=1.3.8
gradio==3.4.1
invisible-watermark
numpy~=1.23.3
omegaconf~=2.2.3
piexif~=1.1.3
Pillow~=9.2.0
pytorch_lightning
realesrgan~=0.3.0
scikit-image>=0.19
timm~=0.6.7
transformers==4.19.2
torch~=1.12.1+cu113
einops~=0.4.1
jsonmerge~=1.8.0
clean-fid
resize-right
torchdiffeq~=0.2.3
kornia~=0.6.7
lark~=1.1.2
flask~=2.2.2
flask-socketio
bunch~=1.0.1
waitress
click~=8.1.3
click-shell
ascii_magic
terminal-img
yachalk~=0.1.5
munch~=2.5.0
tqdm~=4.64.1
torchvision~=0.13.1+cu113
setuptools~=63.2.0
scipy~=1.9.2
clip~=1.0
PyYAML~=6.0
packaging~=21.3
requests~=2.25.1
lmdb~=1.3.0
lpips~=0.1.4
gdown~=4.5.1
facexlib~=0.2.5
opencv-python